requirements.txt
python-telegram-bot
google-generativeai
azure-ai-documentintelligence
pypdf
pandas
Pillow
python-dotenv
openpyxl
PyMuPDF
aiohttp
httpx
google-cloud-storage
boto3
pyarrow
fastparquet
google-cloud-aiplatform
PyMuPDF